requirements.txt
aiohttp
python-socketio
pyyaml